organize componenets navigators and screens
