+++ README.md
@@ -5,16 +5,19 @@
 [expo CLI](https://docs.expo.dev/more/expo-cli/)로 앱 빌드 및 라이브러리 설치
 
 프로젝트 dependency 설치
+
 ```bash
 npx expo install
 ```
 
 앱 빌드
+
 ```bash
 npx expo start
 ```
 
 thrid-party 라이브러리 설치
+
 ```bash
 npx install 패키지
 ```
@@ -22,4 +25,5 @@ npx install 패키지
 ## 코드 설명
 
 ### App.js
+
 앱 entrypoint(앱 누르면 제일 먼저 실행되는 파일)
