--- app/screens/index.ts
@@ -1,7 +1,0 @@
-export * from "./LoginScreen"
-export * from './SignUpScreen'
-export * from './SetProfileScreen'
-export * from './MainScreen'
-export * from './MyFeedScreen'
-export * from './CameraScreen'
-export * from './StudyWithMeListScreen'
