+++ components/AutoImage.tsx
@@ -1,16 +1,16 @@
-import React, { useLayoutEffect, useState } from "react"
-import { Image, ImageProps, ImageURISource, Platform } from "react-native"
+import React, { useLayoutEffect, useState } from "react";
+import { Image, ImageProps, ImageURISource, Platform } from "react-native";
 
 // TODO: document new props
 export interface AutoImageProps extends ImageProps {
   /**
    * How wide should the image be?
    */
-  maxWidth?: number
+  maxWidth?: number;
   /**
    * How tall should the image be?
    */
-  maxHeight?: number
+  maxHeight?: number;
 }
 
 /**
@@ -27,27 +27,32 @@ export function useAutoImage(
   remoteUri: string,
   dimensions?: [maxWidth?: number, maxHeight?: number],
 ): [width: number, height: number] {
-  const [[remoteWidth, remoteHeight], setRemoteImageDimensions] = useState([0, 0])
-  const remoteAspectRatio = remoteWidth / remoteHeight
-  const [maxWidth, maxHeight] = dimensions ?? []
+  const [[remoteWidth, remoteHeight], setRemoteImageDimensions] = useState([
+    0, 0,
+  ]);
+  const remoteAspectRatio = remoteWidth / remoteHeight;
+  const [maxWidth, maxHeight] = dimensions ?? [];
 
   useLayoutEffect(() => {
-    if (!remoteUri) return
+    if (!remoteUri) return;
 
-    Image.getSize(remoteUri, (w, h) => setRemoteImageDimensions([w, h]))
-  }, [remoteUri])
+    Image.getSize(remoteUri, (w, h) => setRemoteImageDimensions([w, h]));
+  }, [remoteUri]);
 
-  if (Number.isNaN(remoteAspectRatio)) return [0, 0]
+  if (Number.isNaN(remoteAspectRatio)) return [0, 0];
 
   if (maxWidth && maxHeight) {
-    const aspectRatio = Math.min(maxWidth / remoteWidth, maxHeight / remoteHeight)
-    return [remoteWidth * aspectRatio, remoteHeight * aspectRatio]
+    const aspectRatio = Math.min(
+      maxWidth / remoteWidth,
+      maxHeight / remoteHeight,
+    );
+    return [remoteWidth * aspectRatio, remoteHeight * aspectRatio];
   } else if (maxWidth) {
-    return [maxWidth, maxWidth / remoteAspectRatio]
+    return [maxWidth, maxWidth / remoteAspectRatio];
   } else if (maxHeight) {
-    return [maxHeight * remoteAspectRatio, maxHeight]
+    return [maxHeight * remoteAspectRatio, maxHeight];
   } else {
-    return [remoteWidth, remoteHeight]
+    return [remoteWidth, remoteHeight];
   }
 }
 
@@ -57,8 +62,8 @@ export function useAutoImage(
  * - [Documentation and Examples](https://github.com/infinitered/ignite/blob/master/docs/Components-AutoImage.md)
  */
 export function AutoImage(props: AutoImageProps) {
-  const { maxWidth, maxHeight, ...ImageProps } = props
-  const source = props.source as ImageURISource
+  const { maxWidth, maxHeight, ...ImageProps } = props;
+  const source = props.source as ImageURISource;
 
   const [width, height] = useAutoImage(
     Platform.select({
@@ -66,7 +71,7 @@ export function AutoImage(props: AutoImageProps) {
       default: source?.uri as string,
     }),
     [maxWidth, maxHeight],
-  )
+  );
 
-  return <Image {...ImageProps} style={[{ width, height }, props.style]} />
+  return <Image {...ImageProps} style={[{ width, height }, props.style]} />;
 }
