+++ models/AuthenticationStore.ts
@@ -1,0 +1,41 @@
+import { Instance, SnapshotOut, types } from "mobx-state-tree";
+
+export const AuthenticationStoreModel = types
+  .model("AuthenticationStore")
+  .props({
+    authToken: types.maybe(types.string),
+    authEmail: "",
+    authPassword: "",
+  })
+  .views((store) => ({
+    get isAuthenticated() {
+      return !!store.authToken;
+    },
+    get validationError() {
+      if (store.authEmail.length === 0) return "can't be blank";
+      if (store.authEmail.length < 6) return "must be at least 6 characters";
+      if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(store.authEmail))
+        return "must be a valid email address";
+      return "";
+    },
+  }))
+  .actions((store) => ({
+    setAuthToken(value: string) {
+      store.authToken = value;
+    },
+    setAuthEmail(value: string) {
+      store.authEmail = value.replace(/ /g, "");
+    },
+    setAuthPassword(value: string) {
+      store.authPassword = value;
+    },
+    logout() {
+      store.authToken = undefined;
+      store.authEmail = "";
+    },
+  }));
+
+export interface AuthenticationStore
+  extends Instance<typeof AuthenticationStoreModel> {}
+export interface AuthenticationStoreSnapshot
+  extends SnapshotOut<typeof AuthenticationStoreModel> {}
