+++ navigators/index.ts
@@ -1,0 +1,2 @@
+export * from "./AppNavigator";
+export * from "./TabNavigator";
