+++ screens/MainScreen.tsx
@@ -1,0 +1,8 @@
+import React, { FC } from "react";
+import { BasicScreen } from "../components/BasicScreen";
+
+export function MainScreen(_props) {
+  const { navigation } = _props;
+
+  return <BasicScreen></BasicScreen>;
+}
